More visibility for gateway and tool status and auto healing capabilities (#215)
* changes implemented reachable and enabled as parameters for status of gateways and tools


* Updated test cases as per new changes for status in gateways and tools


* minor changes


* black and isort


* Alembic script for db migration


* offline tools rpc response


* offline tools rpc response message


* minor changes


* minor flake8 issues


* db migration for postgresql


* Check for existing tables in migration script

* db migration from status to enabled and reachable


* reachable value is updated as necessary


* Updated test cases and linting


* Use bootstrap script

* linting


* Use bootstrap script in main for table creation

* Revert "Use bootstrap script in main for table creation"

This reverts commit 20fccebce2cc6b3c316fbc0e8326f3a77fc2c743.

* Revert "linting"

This reverts commit 2ecc431cce99d5322fb96938fa1d4afbb0cef2a5.

* Revert "Use bootstrap script"

This reverts commit d579e8ee1f147e125c72a3a9095c17d88424c7d0.

---------

+++ alembic/versions/e4fc04d1a442_add_annotations_to_tables.py
@@ -14,6 +14,7 @@ from alembic import op
 
 # Third-Party
 import sqlalchemy as sa
+from sqlalchemy.orm import Session
 
 # revision identifiers, used by Alembic.
 revision: str = 'e4fc04d1a442'
@@ -30,6 +31,14 @@ def upgrade() -> None:
     table. It includes a server-side default of an empty JSON object ('{}') to ensure
     that existing rows get a non-null default value.
     """
+    bind = op.get_bind()
+    sess = Session(bind=bind)
+    inspector = sa.inspect(bind)
+
+    if not inspector.has_table("gateways"):
+        print("Fresh database detected. Skipping migration.")
+        return
+
     op.add_column('tools', sa.Column('annotations', sa.JSON(), server_default=sa.text("'{}'"), nullable=False))
 
 
@@ -40,4 +49,12 @@ def downgrade() -> None:
     This function provides a way to undo the migration, safely removing the
     'annotations' column from the 'tool' table.
     """
+    bind = op.get_bind()
+    sess = Session(bind=bind)
+    inspector = sa.inspect(bind)
+    
+    if not inspector.has_table("gateways"):
+        print("Fresh database detected. Skipping migration.")
+        return
+    
     op.drop_column('tools', 'annotations')
